// src/app/api/models/room-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RoomOwnerResponseDto } from '../models/room-owner-response-dto';
export interface RoomResponseDto {
  capacity?: number;
  createdAt: string;
  description?: string;
  id: number;
  isActive: boolean;
  location?: string;
  name: string;
  owner: RoomOwnerResponseDto;
  updatedAt: string;
}