// src/app/api/models/update-room-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateRoomDto {

  /**
   * Room capacity
   */
  capacity?: number;

  /**
   * Room description
   */
  description?: string;

  /**
   * Room active status
   */
  isActive?: boolean;

  /**
   * Room location
   */
  location?: string;

  /**
   * Room name
   */
  name?: string;
}